+++ gframe/replay.h
@@ -95,7 +95,7 @@ public:
 		uint32_t start_lp;
 		uint32_t start_hand;
 		uint32_t draw_count;
-		uint32_t duel_flags;
+		uint64_t duel_flags;
 	};
 	duel_parameters params;
 	std::string scriptname;
