+++ gframe/single_mode.cpp
@@ -55,7 +55,7 @@ void SingleMode::SetResponse(void* resp, uint32_t len) {
 }
 int SingleMode::SinglePlayThread(DuelOptions duelOptions) {
 	Utils::SetThreadName("SinglePlay");
-	uint32_t opt = duelOptions.duelFlags;
+	uint64_t opt = duelOptions.duelFlags;
 	std::string script_name = "";
 	auto InitReplay = [&]() {
 		uint16_t buffer[20];
